Author is a Post attribute (closes #9)
In my tests I've found that most feeds do not provide an author in their
entries, so Im leaving this pretty much 'as is'. If the entry does
provide an author we will use it, otherwise we will just stick with the
author specified on planet.yml for the blog.