Use `reverse_markdown` instead of `sanitize_html`
Using `sanitize_html` was causing formatting issues on imported content with
non-trivial formatting. For example, simply dropping tags caused blocks of the
form
<noscript>
    some blah blah blah
</noscript>
to be treated and formatted as code blocks, because markdown causes indented
content to be parsed as a code block. Similarly, when a post's entire content is
stored on a single line in the RSS feed, this causes unusual rendering side
effects when tags are deleted, again due to Markdown's convention of treating
continuous blocks of text as paragraphs.

Instead, we use `reverse_markdown` to convert the HTML to Markdown, thereby
preserving the content one cares about (bold/italic, images, tables, lists,
links, etc.), while stripping out the undesirable content (scripts, CSS, etc.).
This has so far proven to be more reliable than the previous HTML sanitization.
It has the added benefit that the entire site, including the aggregated content,
is styled identically.